feat(help-page): add mobile version
